refactor: separate out return and cancellation from status updates
